drivers: digital-io: max149x6 : Add base driver support for max149x6
Created folder with initial source and header file for the max149x6
chip family.
Moved register read, write and update function from the max14906 driver
to the max149xx driver.
Also moved the device descriptor as well as the init_param from the max14906
driver to the max149xx driver, without changing their name so it doesn't
affect te initial max14906 driver's functionality.
Moved the common register used for register writing and reading from the
max14906 driver to the 149x6 driver.
